Include database config name in performance bar
As part of https://gitlab.com/groups/gitlab-org/-/epics/5759 we're
preparing the GitLab application to handle connections to multiple
databases. As such we want to improve our observability to understand
these multiple databases.

This MR just adds the database name to the performance bar details.

Unfortunately the default rails database config name is `primary`.
Eventually we will force this to be `main` and it will be distinct from
the new `ci` database. This does create a little bit of confusion now as
we also use the words `primary/replica` to distinguish between our
read/write and read-only database servers. This means you may see
"Replica" next to "Config name: primary" which could be confusing. I
hope that the usage of "Config name" should help to disambiguate. I have
also used a feature flag to be less disruptive.