Add pages_file_entries to plan_limits
Adds file entries limit to pages deployments:
https://gitlab.com/gitlab-org/gitlab/-/issues/320793.

GitLab Pages daemon actually caches index of each archive in memory,
so archives with too many entries can increase the memory usage.

To prevent this, we adding the limit. It also will work for
self-managed,  but we provide a workaround to change it in docs.

Changelog: added