Fix tag ref detection for pipelines
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/288049

**Problem**

We incorrectly run pipelines for protected tags when a branch with the
same name as tag's name is created.

**Reason**

We use `#tag_exists?` method to identify if current ref is a tag. But
it is wrong, because we use for that a short ref version.

For example, there are branch and tag with the same name:
`refs/heads/test` and `refs/tags/test`.

When we create a pipeline for the branch `refs/heads/test`,
`#tag_exists?` will return true anyway. It takes name `test` and
successfully finds a tag with this name.

**Solution**

Additionally verify if `origin_ref` is a tag

Changelog: security