Use a single lock in PartitionManager
The previous code was using try_obtain once and two seperate locks,
so concurrent code would skip the critical section.

Switching to using in_lock and a single lock ensures that all concurrent
runners get to execute the code, and that partition creation and
partition detaching do not interfere with each other.