Move "Show latest version" button to the left
The buttons at the top of the Merge Request diff view are a bit all over
the place. This change is one step to make them be more intuitive, as
part of https://gitlab.com/gitlab-org/gitlab/-/issues/330579.

The "Show latest version" button was a bit out of place being between
the diff stats and diff preferences.

This change moves that button to the left. It now sits right next to the
commit Prev/Next buttons or version dropdowns. This brings the buttons
which control the range of the diff together.

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/330579
Changelog: changed