Fix scan execution policy selection
This cleans up the policy selection handling by directly storing the
selected policy object instead of computing it based on the policy'
name. This fixes a bug where the policy drawer would remain empty when
selecting a scan execution policy as the `selectedPolicy` object was
being computed based on `networkPolicies` only.

This also fixes a bug where editing a network policy would lead to an
error 404. Including the policy kind in the edit URL should resolve the
issue as the backend now expects this parameter to load policy data to
be passed to the frontend.

Changelog: fixed
EE: true