Remove timeago methods exposed on window object
We currently expose `window.timeago` and `window.gl.utils.localTimeAgo`.
As far as I can tell, they are not used like that anymore and any part
of the code base that still uses it like that should move to direct
imports.

This was actually the _last_ instance of a JavaScript file setting up a
`window.gl` side-effect. Unfortunately this breaks a lot of tests.
Therefore we add the `window.gl` sideeffect to our jest set up. And in
order to be 100% sure that we do not break production, we add the
sideeffect to main.js as well.