git: Remove Repository#rebase_in_progress?
In [1] the implementation of `UserRebaseConfirmable` was rewritten in
Go. This rewrite no longer uses worktrees. Because of this, Gitaly is
pretty much stateless in regard of a rebase. Also in [2] some work was
done to keep the status of a rebase completely in the model, and no
longer ask Gitaly for the status.

This change removes all calls of `IsRebaseInProgress` and makes the code
completely rely on internal status kept in the `MergeRequest` model.

1. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2876
2. d31b733fee8 (Only read rebase status from the model, 2019-08-15)

Issue: https://gitlab.com/gitlab-org/gitaly/-/issues/3614