Update GraphQL project type field container_registry_enabled
Update the field to check container registry visibility. So if the
container registry is enabled, but the API user doesn't have
permission for it, container_registry_enabled in the GraphQL API will
return false.

This is because we are migrating towards using
container_registry_access_level which allows for different users to
have different visibility of the container registry.

Changelog: changed