Fix public selection of pages permissions
The MR https://gitlab.com/gitlab-org/gitlab/-/merge_requests/65723
fixed a bug where the incorrect visibility was stored in the audit
logs.

However, this caused a regression that broke the permissions
for private projects with public pages.
This change reverts the change to avoid it being included in the
14.1 release. It does retain the newly added specs though.

Changelog: fixed