Fix spam checking to consider updater of issue instead of author
When an issue or some other issauble is updated, we previously would
check whether the original author was the support bot instead of the
user who updated this issue. As a result, the spam check would be
triggered unnecessarily and causing spam logs to be attributed to the
support bot.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/335971

Changelog: fixed