Prevent terms from being created if blank
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/65007
introduced the ability to remove t&cs in the admin panel
but this had the unintended consequence of removing
the terms whenever any other application
setting was changed.

This commit reverts the changes to avoid this problem
but does reintroduce the problem that admins can no longer
remove terms and conditions.

Changelog: fixed