Add error tracking client key
To authenticate error tracking events in collector we need to make sure
request has the same public key as the one project has. Keys are safe to
to keep public because they only allow submission of new events.
We do not allow read access to any information for those keys.
So it should not be used anywhere else except error tracking collector.

Signed-off-by: Dmitriy Zaporozhets <dmitriy.zaporozhets@gmail.com>

Changelog: added
Signed-off-by: Dmitriy Zaporozhets <dmitriy.zaporozhets@gmail.com>