Remove unused method InstanceSecurityDashboard#all_pipelines
This method is doing a cross-join between `ci_*` and non `ci_*` tables
which will not be possible when we move `ci_*` tables to a separate
database (see https://gitlab.com/groups/gitlab-org/-/epics/6289 ).

Based on a `git grep all_pipelines` and also
`git grep InstanceSecurityDashboard` I cannot seem to find any uses of
this method. In fact we also removed `Namespace#all_pipelines` a while
back too because it was unused
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/62266 .

It appears this method was added in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/23232 but also
trying to search for usages of `all_pipelines` there I still cannot find
anything as it appears it may have just been replacing another method
that was unused.