Remove DOMPurify attribute sanitization
Due to a bug in DOMPurify, we had to implement
potentially dangerous data-* attribute sanitization using
hooks. These checks can be removed now as they are
no longer needed.

This will allow v-safe-html to configure if it should allow
these attributes for inputs that are safe.

This reverts changes introduced in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/65301

Changelog: changed