Handle post processed content scenario for MR discussions cache
We cache MR discussions after post processing. Before this fix, we
aren't considering the post processing result (e.g. reference state
change) which results to stale cache.

This fix adds a `Note#post_processed_cache_key` that is also added
to `Discussion#cache_key`. This way, when the post processing
changes the note's content for a certain user, we will be able to
show updated content.