Improves find function virtual scrolling disabling
Changes the key code to be `mod+f` to make sure `ctrl+f`
isn't captured on mac OS.
Adds a `mod+g` to make sure it works fully on firefox.
Adds a blur event so that we can be pretty certain that the
user is focused on the find box.