Use parent integration settings when reverting from custom settings
When switching from custom settings to inherited settings, we rely on
the frontend to fill in the inherited values. For encrypted fields we
can't do that, so if one of them is required it's not possible to revert
an integration from custom settings back to default settings.

This is fixed by ignoring the settings sent by the frontend and instead
propagating the values from the parent integration.

Changelog: fixed