Replace custom sorting with GlSorting component
A snapshot test was removed, since it provided almost no value.

Using GlSorting has changed the styling slightly, although not in
a harmful way: the width of the dropdown is fixed, instead of scaling
with the amount of text in the selected item.

The GlSorting component may not be fully Pajamas-compliant, and fixing
that may bring this UI back to nearly how it looked previously.

Addresses part of https://gitlab.com/gitlab-org/gitlab/-/issues/335066.