Connect Puma low-level error handler to Sentry in puma.rb example
We saw in
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/5194 that
errors in a middleware would only be caught by Puma's low-level
handler. This error only showed up in the Puma stderr log, which made it
difficult to diagnose.

If Sentry is configured, we send the exception to Sentry. The error
still will get logged on stderr by Puma.

This mirrors the change we're making in CNG and Omnibus:

1. https://gitlab.com/gitlab-org/charts/gitlab/-/issues/2818
2. https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/6284

Testing notes are in
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/5490.