Remove build trace section related application code
Remove the application code that populates the ci_build_trace_sections
and ci_build_trace_section_names. This data is not read by the
application currently and we can save on space/maintaince and solve
problems with background migration for self-managed users by dropping
the table is subsequent MRs.
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/66674

This table was originally created to collect metrics on jobs but was
never used and an expiration policy of was never enforced.
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/2505

There is the possibility that we will need to build out features that
need this data in the future, based on the future vision, but at this
time it is not being used, and we can re-add it while re-considering
the architecture at that time (likely in 2022).
https://gitlab.com/gitlab-org/gitlab/-/issues/32565