Disable browserslist warning in CI jobs
browserslist throws a warning if certain dependencies are too old. As we
generally are not backporting browserslist updates, this can lead to
errors in CI. Previously we have caught the error in rake, but they have
changed the error message so it is failing again.

By simply setting the BROWSERSLIST_IGNORE_OLD_DATA env variable, we can
circumvent this issue.