Convert setting the full path via SetFullPath RPC
When writing a repo's path (e.g. "gitlab-org/gitlab") into the
gitconfig, we currently use the `SetConfig()` RPC. This RPC has been
deprecated by Gitaly in favor of a new `SetFullPath()`, which can only
be used to explicitly write the repo's path into the gitconfig.

Convert the code to use the new RPC.