Remove caching of Connection#config
Caching was added in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/65262, which lead
to an incident and support for using uncached configs in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/66988.

In this commit we remove caching of the config settings, and add a test
and comment to explain why we don't want the data to be cached. The
performance impact is likely not worth potentially breaking code that
expects uncached settings.

This fixes https://gitlab.com/gitlab-org/gitlab/-/issues/337054.