Add source instance version validation for project
The Project Migration feature requires the source GitLab
instance to be a certain minimal version. If this version is
not met we have to inform the user that the Project Migraiton
might fail or not migrate the whole data.

This commit adds the version control on the page where we show
portable groups. The version validation is now also sent to the
FE so it know when to show the message.

Changelog: added