Add request urgency to logs and metrics
This adds the request urgency to logs by adding a `request_urgency`
and a `request_urgency_s` fields. The duration in seconds will make it
easier to build dashboards showing SLIs that don't meet SLO without
having to copy the information in the runbooks.

The way fields are added to logs is slighty refacotred and tests
added extended.

The `rails_request_apdex` metrics get initialized and emitted with an
`urgency` label.