Improve Marginalia line filtering
In development mode, Marginalia comments include the line that
originated the request in the SQL development log. However, the default
filter fails to ignore the load balancing code, so the comment would
usually attribute the call to the connection proxy.

We now filter out some lines to help narrow in the real caller of the
SQL query.