Wrap Sidekiq scheduler threads in Rails reloader
This backports https://gitlab.com/gitlab-org/gitlab/-/merge_requests/73716
to the 14-4-stable branch.

As a result of enabling load balancing by default in 14.4, we execute DB
queries in our Sidekiq client middleware. This commit is needed to
properly checkin DB connections after every poll. On the next poll, a
new connection will be checked out and a connection check query will be
executed. Rails would then handle the reconnection for us in case the
connection is bad.  Without this, we would continue using a bad
connection until Sidekiq is restarted.